# contracts/openapi/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "GruaHub API",
    "description" : "Plataforma multitenant para gestão de máquinas de pelúcia",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ],
...
  "paths" : {
...
    "/api/v1/payments" : {
      "get" : {
        "tags" : [ "Payment Resource" ],
        "parameters" : [ {
          "name" : "machineId",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "default" : 50,
            "type" : "integer"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponsePaymentResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "PLATFORM_ADMIN", "TENANT_ADMIN", "FINANCE" ]
        } ]
      }
    },
    "/api/v1/payments/sandbox/confirm/{transactionId}" : {
      "post" : {
        "tags" : [ "Payments" ],
        "summary" : "[SANDBOX] Confirmar pagamento fictício",
        "parameters" : [ {
          "name" : "transactionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "X-Machine-Id",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "X-Tenant-Id",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/payments/sandbox/fail/{transactionId}" : {
      "post" : {
        "tags" : [ "Payments" ],
        "summary" : "[SANDBOX] Falhar pagamento fictício",
        "parameters" : [ {
          "name" : "transactionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/payments/sandbox/initiate" : {
      "post" : {
        "tags" : [ "Payments" ],
        "summary" : "[SANDBOX] Criar transação de pagamento fictícia",
        "parameters" : [ {
          "name" : "X-Machine-Id",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "X-Tenant-Id",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/payments/webhook/{provider}" : {
      "post" : {
        "tags" : [ "Payments" ],
        "summary" : "Receber webhook de pagamento",
        "parameters" : [ {
          "name" : "provider",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Idempotency-Key",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "X-Signature",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "X-Tenant-Id",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "format" : "binary",
                "type" : "string"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/v1/reconciliation" : {
      "get" : {
        "tags" : [ "Reconciliation Resource" ],
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "default" : 50,
            "type" : "integer"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseReconciliationCaseResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "PLATFORM_ADMIN", "TENANT_ADMIN", "FINANCE" ]
        } ]
      }
    },
    "/api/v1/reconciliation/summary" : {
      "get" : {
        "tags" : [ "Reconciliation Resource" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "PLATFORM_ADMIN", "TENANT_ADMIN", "FINANCE" ]
        } ]
      }
    },
    "/api/v1/reconciliation/{id}" : {
      "get" : {
        "tags" : [ "Reconciliation Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReconciliationCaseResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "PLATFORM_ADMIN", "TENANT_ADMIN", "FINANCE" ]
        } ]
      }
    },
    "/api/v1/reconciliation/{id}/resolve" : {
      "post" : {
        "tags" : [ "Reconciliation Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ManualResolveRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReconciliationCaseResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "PLATFORM_ADMIN", "TENANT_ADMIN", "FINANCE" ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Instant" : {
        "format" : "date-time",
        "type" : "string",
        "example" : "2022-03-10T16:15:50Z"
      },
...
      "ManualResolveRequest" : {
        "required" : [ "resolution" ],
        "type" : "object",
        "properties" : {
          "resolution" : {
            "pattern" : "\\S",
            "type" : "string"
          },
          "note" : {
            "type" : "string"
          }
        }
      },
...
      "PageResponsePaymentResponse" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PaymentResponse"
            }
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          }
        }
      },
      "PageResponseReconciliationCaseResponse" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReconciliationCaseResponse"
            }
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          }
        }
      },
...
      "PaymentResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "providerTransactionId" : {
            "type" : "string"
          },
          "provider" : {
            "type" : "string"
          },
          "amountCents" : {
            "format" : "int64",
            "type" : "integer"
          },
          "currency" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "machineId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "machineAssetNumber" : {
            "type" : "string"
          },
          "paymentMethod" : {
            "type" : "string"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "confirmedAt" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
      "ReconciliationCaseResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "paymentTransactionId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "creditGrantId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "playSessionId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "status" : {
            "type" : "string"
          },
          "statusReason" : {
            "type" : "string"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "resolvedAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "resolvedBy" : {
            "type" : "string"
          }
        }
      },
...
      "UUID" : {
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
        "type" : "string"
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "openIdConnect",
        "description" : "Authentication",
        "openIdConnectUrl" : "http://localhost:8180/realms/gruahub/.well-known/openid-configuration"
      }
    }
  }
}